requirements.txt
discord.py~=1.5.1
python-dotenv~=0.15.0
aiohttp~=3.6.3
pyrebase4~=4.4.1